requirements.txt
# -------------------------
# Core Gmail + Google Cloud
# -------------------------
google-api-python-client==2.127.0
google-auth==2.34.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-cloud-firestore==2.16.0

# -------------------------
# Vector DB (Chroma)
# -------------------------
chromadb==0.5.5
pinecone
# -------------------------
# Embeddings
# -------------------------
sentence-transformers==2.7.0
torch==2.3.0
numpy==1.26.4

# -------------------------
# LLM Clients
# -------------------------
groq==0.9.0
  # Optional fallback (not required — remove if not needed)

# -------------------------
# Backend API
# -------------------------
fastapi==0.115.0
uvicorn==0.30.6

# -------------------------
# Agent Framework
# -------------------------
langgraph
langchain
# langchain-community
# langchain-core

# -------------------------
# Utilities
# -------------------------
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
chardet==5.2.0
tqdm==4.66.4

# -------------------------
# UI (Optional)
# -------------------------
streamlit==1.38.0

